F and NUM layer changed
Trying to have NUM layer key on home row on g key.
Moved quote and tilde key as well to new position on right hand
side so I don't have to reach as much when typing.
Marking on layout which key activated layer to avoid wierd hand
and finger positions to reach keys on layer