screenshots and rotary encoder
added shortcut to take screenshot either saving
it file or clipboard.
Trying to use the right rotary encoder to navigate
back and forward in Intellij.
Having `v` and `m` as layer keys wasn't really
working, so moved to `d` and `k`. Usung `v` and
`m` was interferring with other shortcuts.