layer keys update
switch placement of layer key for NAV and NUM. I shouldn't move my finger for
numbers, but when I use arrow keys I am not in typing mode so staying on
home row is less important.

Shift keys were not working for selecting stuff.

Layer key for F keys has been changed to MO function as I found that the single
tap wasn't working when I wanted to press same F key multiple times.

Experimenting with removing RGUI and replacing it with | \ key. The RGUI is
now where the \ | used to be.
Experimenting with lead key timeout, as getting the lead especially to work
wasn't working too great